requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-debug-toolbar==1.4
django-filter==0.12.0
django-toolbelt==0.0.1
django-webpack-loader==0.3.0
djangorestframework==3.3.2
gunicorn==19.4.5
heroku==0.1.4
optional-django==0.3.0
psycopg2==2.6.1
python-dateutil==1.5
requests==2.9.1
sqlparse==0.1.18
static3==0.6.1
whitenoise==2.0.6
